+++ test/solana_wallet_provider_test.dart
@@ -1,4 +1,2 @@
-import 'package:flutter_test/flutter_test.dart';
-import 'package:solana_wallet_provider/solana_wallet_provider.dart';
 
 void main() {}